Move 'Send Reminder' to partial
The 'Send Reminder' button is better suited to a partial rather than a
decorator.  This makes it easier to include other components as well,
such as the checkbox and the message for the checkbox.  It also allows
for DRYing up the `views.en.yml` a little, since the relevant title and
tooltip no longer have to be in two separate spots.